Do not user search criteria in IndexService
